Point to events.js to get the event classes now and moved the custom Range interface into there.
